scripts/requirements.txt
pandas==2.0.0
tabula-py==2.7.0
pdfplumber==0.9.0
camelot-py[cv]==0.11.0
openpyxl==3.1.2
ghostscript==0.7
opencv-python==4.8.0.74
pdf2image==1.16.3 